% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_selected}
\alias{delete_selected}
\title{Delete selected elements from a tsjson object}
\usage{
delete_selected(json)
}
\arguments{
\item{json}{tsjson object.}
}
\value{
Modified tsjson object.
}
\description{
TODO
}
