% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{deleted}
\alias{deleted}
\title{Special marker to delete elements from a tsjson object}
\usage{
deleted()
}
\value{
A marker object to be used at the right hand side of a tsjson
"selection assignment", see examples below.
}
\description{
TODO
}
\examples{
# TODO
}
