% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format.tsjson}
\alias{format.tsjson}
\title{Format a tsjson object}
\usage{
\method{format}{tsjson}(x, n = 10, ...)
}
\arguments{
\item{x}{tsjson object.}

\item{n}{Number of lines, or number of selections to print.}

\item{...}{Ignored.}
}
\value{
Character vector.
}
\description{
TODO
}
