% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-json.R
\name{format_selected}
\alias{format_selected}
\title{Format the selected JSON elements}
\usage{
format_selected(
  json,
  format = c("pretty", "oneline", "compact"),
  options = NULL
)
}
\arguments{
\item{json}{tsjson object.}

\item{format}{Formatting, one of:
\itemize{
\item \code{"pretty"}: arrays and objects are formatted in multiple lines,
\item \code{"compact"}: format everything without whitespace,
\item \code{"oneline"}: format everything without newlines, but include
whitespace after commas, colons, opening brackets and braces, and
before closing brackets and braces.
}}

\item{options}{List of options, see \code{\link[=tsjson_options]{tsjson_options()}}. This argument
must be named and cannot be abbreviated.}
}
\value{
The updated tsjson object.
}
\description{
Format the selected JSON elements
}
\details{
If \code{json} does not have a selection, then all of it is formatted.
If \code{json} has an empty selection, then nothing happens.
}
\examples{
json <- load_json(text = "{ \"a\": [1,2,3] }")
json

json |> format_selected()

json |> select("a") |> format_selected()
}
