% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-into-selections.R
\name{insert_into_selected}
\alias{insert_into_selected}
\title{Insert a new element into the selected ones in a tsjson object}
\usage{
insert_into_selected(
  json,
  new,
  key = NULL,
  at = Inf,
  format = c("auto", "pretty", "compact", "oneline")
)
}
\arguments{
\item{json}{tsjson object}

\item{new}{New element to insert. Will be serialized with
\code{\link[=serialize_json]{serialize_json()}}.}

\item{key}{Key of the new element, when inserting into an object.}

\item{at}{What position to insert the new element at:
\itemize{
\item \code{0}: at the beginning,
\item \code{Inf}: at the end,
\item other numbers: after the specified element.
}}

\item{Formatting}{of the \code{new} element, passed to \code{\link[=serialize_json]{serialize_json()}}.}
}
\value{
The modified tsjson object.
}
\description{
TODO
}
