% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-json.R
\name{load_json}
\alias{load_json}
\title{Parse a JSON file or string into a tsjson object}
\usage{
load_json(file = NULL, text = NULL, ranges = NULL, options = NULL)
}
\arguments{
\item{file}{Path of a JSON file. Use either \code{file} or \code{text}.}

\item{text}{JSON string. Use either \code{file} or \code{text}.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}

\item{options}{List of options, see \code{\link[=tsjson_options]{tsjson_options()}}. This argument
must be named and cannot be abbreviated.}
}
\value{
A tsjson object.
}
\description{
Parse a JSON file or string and create a tsjson object that represents
its document. This object can then be queried and manipulated.
}
\details{
tsjson objects have \code{\link[=format.tsjson]{format()}} and
\code{\link[=print.tsjson]{print()}} methods to pretty-print them to the screen.

They can be converted to a data frame using the
\link[=tsjson-brackets]{single bracket} operator.
}
\examples{
json <- load_json(text = serialize_json(list(
  a = list(a1 = list(1,2,3), a2 = "string"),
  b = list(4, 5, 6),
  c = list(c1 = list("a", "b"))
)))
json
}
\seealso{
\code{\link[=select]{select()}} to select part(s) of a tsjson object,
\code{\link[=unserialize_selected]{unserialize_selected()}} to extract the selected part(s),
\code{\link[=format_selected]{format_selected()}} to format the selected part(s),
\code{\link[=delete_selected]{delete_selected()}}, \code{\link[=insert_into_selected]{insert_into_selected()}} and
\code{\link[=update_selected]{update_selected()}} to manipulate it. \code{\link[=save_json]{save_json()}} to save
the JSON document to a file.
}
