% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tsjson}
\alias{print.tsjson}
\title{Print a tsjson object}
\usage{
\method{print}{tsjson}(x, n = 10, ...)
}
\arguments{
\item{x}{tsjson object.}

\item{n}{Number of lines, or number of selections to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Calls \code{\link[=format.tsjson]{format.tsjson()}} to format the tsjson object, writes the
formatted object to the standard output, and returns the original object
invisibly.
}
\examples{
json <- load_json(text = serialize_json(list(
  a = list(a1 = list(1,2,3), a2 = "string"),
  b = list(4, 5, 6),
  c = list(c1 = list("a", "b"))
)))

json
}
