% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sitter.R
\name{query_json}
\alias{query_json}
\title{Run tree-sitter queries on a JSON file or string}
\usage{
query_json(file = NULL, text = NULL, query, ranges = NULL)
}
\arguments{
\item{file}{Path of a JSON file. Use either \code{file} or \code{text}.}

\item{text}{JSON string. Use either \code{file} or \code{text}.}

\item{query}{Character string, the tree-sitter query to run.}

\item{ranges}{Optionally a list of integer pairs defining ranges within
\code{file} or \code{text} to parse.}
}
\value{
A list with entries \code{patterns} and \code{matched_captures}.
TODO
}
\description{
TODO
}
