% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsjson-package.R
\name{quickstart}
\alias{quickstart}
\title{tsjson quickstart}
\description{
tsjson quickstart
}
\details{
\subsection{Create a tjson object}{

Create a tsjson object from a string:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{getwd()
txt <- r"(
// this is a comment
\{
  "a": \{
    "a1": [1, 2, 3],
    // comment
    "a2": "string"
  \},
  "b": [
    \{
      "b11": true,
      "b12": false
    \},
    \{
      "b21": false,
      "b22": false
    \}
  ]
\}
)"
json <- load_json(text = txt)
}\if{html}{\out{</div>}}

Pretty print a tsjson object:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{json
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># json (19 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span>                                                                           
#> <span style="color: #6F7783;"> 2 | </span>// this is a comment                                                       
#> <span style="color: #6F7783;"> 3 | </span>\{                                                                          
#> <span style="color: #6F7783;"> 4 | </span>  "a": \{                                                                   
#> <span style="color: #6F7783;"> 5 | </span>    "a1": [1, 2, 3],                                                       
#> <span style="color: #6F7783;"> 6 | </span>    // comment                                                             
#> <span style="color: #6F7783;"> 7 | </span>    "a2": "string"                                                         
#> <span style="color: #6F7783;"> 8 | </span>  \},                                                                       
#> <span style="color: #6F7783;"> 9 | </span>  "b": [                                                                   
#> <span style="color: #6F7783;">10 | </span>    \{                                                                      
#> <span style="color: #6F7783;">ℹ 9 more lines</span>                                                                  
#> <span style="color: #6F7783;">ℹ Use `print(n = ...)` to see more lines</span>                                        
</pre></div>
}}

}

\subsection{Select elements in a tsjson object}{

Select element by objects key:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{select(json, "a")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># json (19 lines, 1 selected element)</span>                                           
#>   <span style="color: #6F7783;"> 1  | </span>                                                                        
#>   <span style="color: #6F7783;"> 2  | </span>// this is a comment                                                    
#>   <span style="color: #6F7783;"> 3  | </span>\{                                                                       
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 4  | </span>  "a": <span style="color: #66C2CD;">\{</span>                                                                
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 5  | </span><span style="color: #66C2CD;">    "a1": [1, 2, 3],</span>                                                    
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 6  | </span><span style="color: #66C2CD;">    // comment</span>                                                          
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 7  | </span><span style="color: #66C2CD;">    "a2": "string"</span>                                                      
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 8  | </span><span style="color: #66C2CD;">  \}</span>,                                                                    
#>   <span style="color: #6F7783;"> 9  | </span>  "b": [                                                                
#>   <span style="color: #6F7783;">10  | </span>    \{                                                                   
#>   <span style="color: #6F7783;">11  | </span>      "b11": true,                                                      
#>   <span style="color: #6F7783;">...</span>                                                                           
</pre></div>
}}


Select element inside element:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{select(json, "a", "a1")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># json (19 lines, 1 selected element)</span>                                           
#>   <span style="color: #6F7783;">2   | </span>// this is a comment                                                    
#>   <span style="color: #6F7783;">3   | </span>\{                                                                       
#>   <span style="color: #6F7783;">4   | </span>  "a": \{                                                                
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;">5   | </span>    "a1": <span style="color: #66C2CD;">[1, 2, 3]</span>,                                                    
#>   <span style="color: #6F7783;">6   | </span>    // comment                                                          
#>   <span style="color: #6F7783;">7   | </span>    "a2": "string"                                                      
#>   <span style="color: #6F7783;">8   | </span>  \},                                                                    
#>   <span style="color: #6F7783;">...</span>                                                                           
</pre></div>
}}


Select element(s) of an array:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{select(json, "a", "a1", 1:2)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># json (19 lines, 2 selected elements)</span>                                          
#>   <span style="color: #6F7783;">2   | </span>// this is a comment                                                    
#>   <span style="color: #6F7783;">3   | </span>\{                                                                       
#>   <span style="color: #6F7783;">4   | </span>  "a": \{                                                                
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;">5   | </span>    "a1": [<span style="color: #66C2CD;">1</span>, <span style="color: #66C2CD;">2</span>, 3],                                                    
#>   <span style="color: #6F7783;">6   | </span>    // comment                                                          
#>   <span style="color: #6F7783;">7   | </span>    "a2": "string"                                                      
#>   <span style="color: #6F7783;">8   | </span>  \},                                                                    
#>   <span style="color: #6F7783;">...</span>                                                                           
</pre></div>
}}


Select multiple keys from an object:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{select(json, "a", c("a1", "a2"))
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># json (19 lines, 2 selected elements)</span>                                          
#>   <span style="color: #6F7783;"> 2  | </span>// this is a comment                                                    
#>   <span style="color: #6F7783;"> 3  | </span>\{                                                                       
#>   <span style="color: #6F7783;"> 4  | </span>  "a": \{                                                                
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 5  | </span>    "a1": <span style="color: #66C2CD;">[1, 2, 3]</span>,                                                    
#>   <span style="color: #6F7783;"> 6  | </span>    // comment                                                          
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 7  | </span>    "a2": <span style="color: #66C2CD;">"string"</span>                                                      
#>   <span style="color: #6F7783;"> 8  | </span>  \},                                                                    
#>   <span style="color: #6F7783;"> 9  | </span>  "b": [                                                                
#>   <span style="color: #6F7783;">10  | </span>    \{                                                                   
#>   <span style="color: #6F7783;">...</span>                                                                           
</pre></div>
}}

}

\subsection{Delete elements}{

Delete selected elements:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{select(json, "a", "a1") |> delete_selected()
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># json (18 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span>                                                                           
#> <span style="color: #6F7783;"> 2 | </span>// this is a comment                                                       
#> <span style="color: #6F7783;"> 3 | </span>\{                                                                          
#> <span style="color: #6F7783;"> 4 | </span>  "a": \{                                                                   
#> <span style="color: #6F7783;"> 5 | </span>    // comment                                                             
#> <span style="color: #6F7783;"> 6 | </span>    "a2": "string"                                                         
#> <span style="color: #6F7783;"> 7 | </span>  \},                                                                       
#> <span style="color: #6F7783;"> 8 | </span>  "b": [                                                                   
#> <span style="color: #6F7783;"> 9 | </span>    \{                                                                      
#> <span style="color: #6F7783;">10 | </span>      "b11": true,                                                         
#> <span style="color: #6F7783;">ℹ 8 more lines</span>                                                                  
#> <span style="color: #6F7783;">ℹ Use `print(n = ...)` to see more lines</span>                                        
</pre></div>
}}

}

\subsection{Insert elements}{

Insert element into an array:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{select(json, "a", "a1") |> insert_into_selected(at = 2, "new")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># json (24 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span>                                                                           
#> <span style="color: #6F7783;"> 2 | </span>// this is a comment                                                       
#> <span style="color: #6F7783;"> 3 | </span>\{                                                                          
#> <span style="color: #6F7783;"> 4 | </span>  "a": \{                                                                   
#> <span style="color: #6F7783;"> 5 | </span>    "a1": [                                                                
#> <span style="color: #6F7783;"> 6 | </span>      1,                                                                   
#> <span style="color: #6F7783;"> 7 | </span>      2,                                                                   
#> <span style="color: #6F7783;"> 8 | </span>      "new",                                                               
#> <span style="color: #6F7783;"> 9 | </span>      3                                                                    
#> <span style="color: #6F7783;">10 | </span>    ],                                                                     
#> <span style="color: #6F7783;">ℹ 14 more lines</span>                                                                 
#> <span style="color: #6F7783;">ℹ Use `print(n = ...)` to see more lines</span>                                        
</pre></div>
}}


Inserting into an array reformats the array.

Insert element into an object, at the specified key:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{select(json, "a") |>
  insert_into_selected(key = "a0", at = 0, list("new", "element"))
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># json (27 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span>                                                                           
#> <span style="color: #6F7783;"> 2 | </span>// this is a comment                                                       
#> <span style="color: #6F7783;"> 3 | </span>\{                                                                          
#> <span style="color: #6F7783;"> 4 | </span>  "a": \{                                                                   
#> <span style="color: #6F7783;"> 5 | </span>    "a0": [                                                                
#> <span style="color: #6F7783;"> 6 | </span>      "new",                                                               
#> <span style="color: #6F7783;"> 7 | </span>      "element"                                                            
#> <span style="color: #6F7783;"> 8 | </span>    ],                                                                     
#> <span style="color: #6F7783;"> 9 | </span>    "a1": [                                                                
#> <span style="color: #6F7783;">10 | </span>      1,                                                                   
#> <span style="color: #6F7783;">ℹ 17 more lines</span>                                                                 
#> <span style="color: #6F7783;">ℹ Use `print(n = ...)` to see more lines</span>                                        
</pre></div>
}}

}

\subsection{Update elements}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{select(json, "a", c("a1", "a2")) |> update_selected("completely new")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># json (19 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span>                                                                           
#> <span style="color: #6F7783;"> 2 | </span>// this is a comment                                                       
#> <span style="color: #6F7783;"> 3 | </span>\{                                                                          
#> <span style="color: #6F7783;"> 4 | </span>  "a": \{                                                                   
#> <span style="color: #6F7783;"> 5 | </span>    "a1": "completely new",                                                
#> <span style="color: #6F7783;"> 6 | </span>    // comment                                                             
#> <span style="color: #6F7783;"> 7 | </span>    "a2": "completely new"                                                 
#> <span style="color: #6F7783;"> 8 | </span>  \},                                                                       
#> <span style="color: #6F7783;"> 9 | </span>  "b": [                                                                   
#> <span style="color: #6F7783;">10 | </span>    \{                                                                      
#> <span style="color: #6F7783;">ℹ 9 more lines</span>                                                                  
#> <span style="color: #6F7783;">ℹ Use `print(n = ...)` to see more lines</span>                                        
</pre></div>
}}

}
}
