% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-json.R
\name{save_json}
\alias{save_json}
\title{Write a tsjson object to a file}
\usage{
save_json(json, file = NULL)
}
\arguments{
\item{json}{tsjson object.}

\item{file}{File or connection to write to. Both binary and text
connections are supported.}
}
\value{
Nothing.
}
\description{
TODO
}
