% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_select<-}
\alias{ts_tree_select<-}
\title{Update selected elements in a ts_tree_jsonc object}
\arguments{
\item{x, json}{ts_tree_jsonc object. Create a ts_tree_jsonc object with
\code{\link[tsitter:ts_tree_new]{tsitter::ts_tree_new()}}.}

\item{i, ...}{Selectors, see \code{\link[tsitter:ts_tree_select]{tsitter::ts_tree_select()}}.}

\item{value}{New value. Will be serialized to JSON with
\code{\link[=ts_serialize_jsonc]{ts_serialize_jsonc()}}.}
}
\value{
The updated ts_tree_jsonc object.
}
\description{
Update the selected elements of a JSON document, using the replacement
function syntax.
}
\examples{
library(tsitter)
json <- ts_parse_jsonc("{}")

json <- json |>
  ts_tree_select("r", "editor.formatOnSave") |>
  ts_tree_update(TRUE)
json

json <- json |>
  ts_tree_select("r", "editor.formatOnSave") |>
  ts_tree_delete()
json

# Insert an array
json <- json |>
  ts_tree_select("foo") |>
  ts_tree_update(1:3)
json

# Update the array at location 2
json |> ts_tree_select("foo", 2) |> ts_tree_update(0)

# Insert at location 2
json |> ts_tree_select("foo") |> ts_tree_insert(0, at = 2)

# Insert at the end of the array with `Inf` as `at`
json |> ts_tree_select("foo") |> ts_tree_insert(0, at = Inf)

# Only the modified elements are reformatted
json <- ts_parse_jsonc('{"foo":[1,2],\n"bar":1}')
json |> ts_tree_select("foo") |> ts_tree_insert(0, at = Inf)

# You can control how those elements are formatted
json |> ts_tree_select("foo") |>
  ts_tree_insert(0, at = Inf, options = list(indent_width = 2))
}
\seealso{
Save the updated ts_tree_jsonc object to a file with
\code{\link[tsitter:ts_tree_write]{tsitter::ts_tree_write()}}.
}
\keyword{internal}
