% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\alias{select_refine}
\title{Select elements in a tsjson object}
\usage{
select(json, ...)

select_refine(json, ...)
}
\arguments{
\item{json}{tsjson object.}

\item{...}{Selectors, see below.}
}
\value{
A tsjson object, potentially with some elements selected.
}
\description{
TODO
}
\details{
\subsection{Selectors}{

TODO
}
}
