% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_query}
\alias{select_query}
\title{Select the nodes matching a tree-sitter query in a tsjson object}
\usage{
select_query(json, query)
}
\arguments{
\item{json}{tsjson object.}

\item{query}{String, a tree-sitter query.}
}
\description{
See https://tree-sitter.github.io/tree-sitter/ on writing tree-sitter
queries. Captured nodes of the TOML document will be selected.
}
\examples{
# A very simple JSON document
txt <- "{ \"a\": 1, \"b\": \"foo\", \"c\": 20 }"

# Take a look at it
load_json(text = txt) |> format_selected()

# Select all pairs where the value is a number and change them to 100
load_json(text = txt) |>
  select_query("((pair value: (number) @num))") |>
  update_selected(100)
}
\seealso{
\code{\link[=query_json]{query_json()}} for running a tree sitter query on text and
obtaining the result.
}
