% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize-json.R
\name{serialize_json}
\alias{serialize_json}
\title{Serialize an R object to JSON}
\usage{
serialize_json(
  obj,
  file = NULL,
  collapse = FALSE,
  format = c("pretty", "compact", "oneline")
)
}
\arguments{
\item{obj}{R object to serialize.}

\item{file}{If not \code{NULL} then the result if written to this file.}

\item{collapse}{If \code{file} is \code{NULL} then whether to return a character
scalar or a character vector.}

\item{format}{Formatting, one of:
\itemize{
\item \code{"pretty"}: arrays and objects are formatted in multiple lines,
\item \code{"compact"}: format everything without whitespace,
\item \code{"oneline"}: format everything without newlines, but include
whitespace after commas, colons, opening brackets and braces, and
before closing brackets and braces.
}}
}
\value{
If \code{file} is \code{NULL} then a character scalar (\code{collapse} = TRUE)
or vector (\code{collapse} = FALSE). If \code{file} is not \code{NULL} then nothing.
}
\description{
TODO
}
