% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sitter.R
\name{sexpr_json}
\alias{sexpr_json}
\title{Show the syntax tree structure of a JSON file or string}
\usage{
sexpr_json(file = NULL, text = NULL, ranges = NULL)
}
\arguments{
\item{file}{Path of a JSON file. Use either \code{file} or \code{text}.}

\item{text}{JSON string. Use either \code{file} or \code{text}.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}
}
\description{
Show the syntax tree structure of a JSON file or string
}
\examples{
sexpr_json(text = "{ \"a\": true, \"b\": [1, 2, 3] }")
}
