% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-json.R
\name{[.tsjson}
\alias{[.tsjson}
\title{Convert a tsjson object to a data frame}
\usage{
\method{[}{tsjson}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{tsjson object.}

\item{i, j:}{indices}

\item{drop:}{Ignored.}
}
\description{
TODO
}
