% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{[[.tsjson}
\alias{[[.tsjson}
\title{Extract elements from a tsjson object}
\usage{
\method{[[}{tsjson}(x, i, ...)
}
\arguments{
\item{x}{tsjson object}

\item{i, ...}{Selectors, see \code{\link[=select]{select()}}.}
}
\description{
TODO
}
