% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-parse-jsonc.R
\name{ts_parse_jsonc}
\alias{ts_parse_jsonc}
\alias{ts_read_jsonc}
\title{Parse a JSON file or string into a ts_tree_jsonc object}
\usage{
ts_parse_jsonc(text, ranges = NULL, fail_on_parse_error = TRUE, options = NULL)

ts_read_jsonc(file, ranges = NULL, fail_on_parse_error = TRUE, options = NULL)
}
\arguments{
\item{text}{\Sexpr[stage=render,results=rd]{tsitter:::doc_insert("tsitter::ts_tree_new_param_text", "tsitter")}}

\item{ranges}{\Sexpr[stage=render,results=rd]{tsitter:::doc_insert("tsitter::ts_tree_new_param_ranges", "tsitter")}}

\item{fail_on_parse_error}{\Sexpr[stage=render,results=rd]{tsitter:::doc_insert("tsitter::ts_tree_new_param_fail_on_parse_error", "tsitter")}}

\item{options}{Named list of parsing options, see
\link[=tsjsonc_options]{tsjsonc options}.}

\item{file}{\Sexpr[stage=render,results=rd]{tsitter:::doc_insert("tsitter::ts_tree_new_param_file", "tsitter")}}
}
\value{
A ts_tree_jsonc object.
}
\description{
Parse a JSON file or string and create a ts_tree_jsonc object that
represents its document. This object can then be queried and
manipulated.
}
\details{
ts_tree_jsonc objects have \code{format()} and \code{print()} methods to pretty-print
them to the screen.

They can be converted to a data frame using the
\link[tsitter:ts_tree-brackets]{single bracket} operator.
}
\examples{
library(tsitter)
text <- '
{
  "a": 1,
  "b": [2, 3, 4],
  "[r]": {
    "this": "setting",
    // A comment!
    "that": true
  }
}
'

# Parse the JSON, allowing comments (i.e. JSONC)
ts_parse_jsonc(text)

# Try to parse the JSON, but comments aren't allowed!
try(ts_parse_jsonc(text, options = list(allow_comments = FALSE)))

# Extract parts of the JSON
ts_parse_jsonc(text) |>
  ts_tree_select("b") |>
  ts_tree_unserialize()
ts_parse_jsonc(text) |>
  ts_tree_select("[r]") |>
  ts_tree_unserialize()
ts_parse_jsonc(text) |>
  ts_tree_select("[r]", "that") |>
  ts_tree_unserialize()

# Use a `list()` combining strings and positional indices when
# arrays are involved
ts_parse_jsonc(text) |>
  ts_tree_select("b", 2) |>
  ts_tree_unserialize()
}
\seealso{
\code{\link[tsitter:ts_tree_select]{tsitter::ts_tree_select()}} to select part(s) of a ts_tree_jsonc
object, \code{\link[tsitter:ts_tree_unserialize]{tsitter::ts_tree_unserialize()}} to extract the selected part(s),
\code{\link[tsitter:ts_tree_format]{tsitter::ts_tree_format()}} to format the selected part(s),
\code{\link[tsitter:ts_tree_delete]{tsitter::ts_tree_delete()}}, \code{\link[tsitter:ts_tree_insert]{tsitter::ts_tree_insert()}} and
\code{\link[tsitter:ts_tree_update]{tsitter::ts_tree_update()}} to manipulate it. \code{\link[tsitter:ts_tree_write]{tsitter::ts_tree_write()}} to save
the JSON document to a file.
}
