% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-serialize-jsonc.R
\name{ts_serialize_jsonc}
\alias{ts_serialize_jsonc}
\title{Serialize an R object to JSON}
\usage{
ts_serialize_jsonc(obj, file = NULL, collapse = FALSE, options = NULL)
}
\arguments{
\item{obj}{R object to serialize.}

\item{file}{If not \code{NULL} then the result if written to this file.}

\item{collapse}{If \code{file} is \code{NULL} then whether to return a character
scalar or a character vector.}

\item{options}{\Sexpr[stage=render,results=rd]{tsitter:::doc_insert("tsitter::ts_tree_format_param_options")}\if{html}{\out{
<p>
}}


See details in the manual of the specific parser.}
}
\value{
If \code{file} is \code{NULL} then a character scalar (\code{collapse} = TRUE)
or vector (\code{collapse} = FALSE). If \code{file} is not \code{NULL} then nothing.
}
\description{
Create JSON from an R object.
Note that this function is not a generic serializer that can represent
any R object in JSON. Also, you cannot expect that \code{\link[=ts_unserialize_jsonc]{ts_unserialize_jsonc()}}
will do the exact inverse of \code{\link[=ts_serialize_jsonc]{ts_serialize_jsonc()}}.
}
\details{
tsjsonc functions \code{\link[tsitter:ts_tree_update]{tsitter::ts_tree_update()}} and \code{\link[tsitter:ts_tree_insert]{tsitter::ts_tree_insert()}} use
\code{\link[=ts_serialize_jsonc]{ts_serialize_jsonc()}} to create new JSON code.

See the examples below on how to create all possible JSON elements with
\code{\link[=ts_serialize_jsonc]{ts_serialize_jsonc()}}.
}
\examples{
library(tsitter)
# null
ts_serialize_jsonc(NULL)

# true, false, use a logical scalar
ts_serialize_jsonc(TRUE)
ts_serialize_jsonc(FALSE)

# strings, use a character scalar
ts_serialize_jsonc("string with escapes: \b \ud020")

# number, use a numeric scalar
ts_serialize_jsonc(42.25)

# array, use an unnamed list, i.e. _not_ an atomic vector
txt <- ts_serialize_jsonc(list(1, 2, 3,"x", "y"))
ts_parse_jsonc(txt)

# empty array
ts_serialize_jsonc(list())

# object, use a named (or partially named) list, i.e. _not_ an atomic vector
txt <- ts_serialize_jsonc(list(a = 1, b = 2))
ts_parse_jsonc(txt)

# empty object, use a named empty list
ts_serialize_jsonc(structure(list(), names = character()))
}
\seealso{
\code{\link[=ts_unserialize_jsonc]{ts_unserialize_jsonc()}} for the opposite.
}
