% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-delete.R
\name{ts_tree_delete.ts_tree_jsonc}
\alias{ts_tree_delete.ts_tree_jsonc}
\title{Delete selected elements from a ts_tree_jsonc object}
\usage{
\method{ts_tree_delete}{ts_tree_jsonc}(tree, ...)
}
\arguments{
\item{tree}{\eval{tsitter:::doc_insert("tsitter::ts_tree_delete_param_tree", "tsjsonc")}}

\item{...}{Reserved for future use.}
}
\value{
\eval{tsitter:::doc_insert("tsitter::ts_tree_delete_return", "tsjsonc")}
}
\description{
Delete selected elements from a ts_tree_jsonc object
}
\details{
\eval{tsitter:::doc_insert("ts_tree_delete_details", "tsjsonc")}

Comments appearing inside the deleted elements are also deleted. Other
comments are left as is.
}
\examples{
library(tsitter)
tree <- ts_parse_jsonc("{ \"a\": //comment\ntrue, \"b\": [1, 2, 3] }")
tree

tree |> ts_tree_select("a")
tree |> ts_tree_select("a") |> ts_tree_delete()
}
