% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-format.R
\name{ts_tree_format.ts_tree_jsonc}
\alias{ts_tree_format.ts_tree_jsonc}
\title{Format the selected JSON elements}
\usage{
\method{ts_tree_format}{ts_tree_jsonc}(tree, options = NULL, ...)
}
\arguments{
\item{tree}{ts_tree_jsonc object.}

\item{options}{Named list of formatting options, see
\link[=tsjsonc_options]{tsjsonc options}.}

\item{...}{Reserved for future use.}
}
\value{
The updated ts_tree_jsonc object.
}
\description{
Format the selected JSON elements
}
\details{
If \code{tree} does not have a selection, then all of it is formatted.
If \code{tree} has an empty selection, then nothing happens.
}
\examples{
library(tsitter)
tree <- ts_parse_jsonc("{ \"a\": [1,2,3] }")
tree

tree |> ts_tree_format()

tree |> ts_tree_select("a") |> ts_tree_format()
}
\keyword{internal}
