% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-insert.R
\name{ts_tree_insert.ts_tree_jsonc}
\alias{ts_tree_insert.ts_tree_jsonc}
\title{Insert a new element into the selected ones in a ts_tree_jsonc object}
\usage{
\method{ts_tree_insert}{ts_tree_jsonc}(tree, new, key = NULL, at = Inf, options = NULL, ...)
}
\arguments{
\item{tree}{ts_tree_jsonc object}

\item{new}{New element to insert. Will be serialized with
\code{\link[=ts_serialize_jsonc]{ts_serialize_jsonc()}}.}

\item{key}{Key of the new element, when inserting into an object.}

\item{at}{What position to insert the new element at:
\itemize{
\item \code{0}: at the beginning,
\item \code{Inf}: at the end,
\item other numbers: after the specified element,
\item a character scalar, the key after which the new element is inserted,
if that key exists, when inserting into an object. If this key does
not exist, then the new element is inserted at the end of the object.
}}

\item{options}{Named list of formatting options, see
\link[=tsjsonc_options]{tsjsonc options}.}

\item{...}{Must be empty currently, reserved for future use.}
}
\value{
The modified ts_tree_jsonc object.
}
\description{
Insert a new element into each selected array or object.
}
\details{
It is an error trying to insert into an element that is not an array and
not an object.
}
\examples{
library(tsitter)
json <- ts_parse_jsonc("{ \"a\": true, \"b\": [1, 2, 3] }")
json

json |> ts_tree_select("b") |> ts_tree_insert("foo", at = 1)
}
\keyword{internal}
