% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-query.R
\name{ts_tree_query.ts_tree_jsonc}
\alias{ts_tree_query.ts_tree_jsonc}
\title{Run tree-sitter queries on JSONC tree-sitter trees}
\usage{
\method{ts_tree_query}{ts_tree_jsonc}(tree, query)
}
\arguments{
\item{tree}{\eval{tsitter:::doc_insert("ts_tree_query_param_tree", "tsjsonc")}}

\item{query}{\eval{tsitter:::doc_insert("ts_tree_query_param_query", "tsjsonc")}}
}
\value{
\eval{tsitter:::doc_insert("ts_tree_query_return", "tsjsonc")}
}
\description{
\eval{tsitter:::doc_insert("ts_tree_query_description", "tsjsonc")}
}
\details{
\eval{tsitter:::doc_insert("ts_tree_query_details", "tsjsonc")}
\eval{tsitter:::doc_extra()}
}
\examples{
# Select all numbers in a JSONC document ------------------------------------
library(tsitter)
json <- tsjsonc::ts_parse_jsonc(
  '{ "a": 1, "b": [10, 20, 30], "c": { "c1": true, "c2": 100 } }'
)
json |> ts_tree_query("(number) @number")
}
