% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_select.ts_tree_jsonc}
\alias{ts_tree_select.ts_tree_jsonc}
\title{Select parts of a JSONC tree-sitter tree}
\usage{
\method{ts_tree_select}{ts_tree_jsonc}(tree, ..., refine = FALSE)
}
\arguments{
\item{tree}{\eval{tsitter:::doc_insert("tsitter::ts_tree_select_param_tree", "tsjsonc")}}

\item{...}{Reserved for future use.}

\item{refine}{\eval{tsitter:::doc_insert("tsitter::ts_tree_select_param_refine", "tsjsonc")}}
}
\value{
\eval{tsitter:::doc_insert("tsitter::ts_tree_select_return", "tsjsonc")}
}
\description{
\eval{tsitter:::doc_insert("tsitter::ts_tree_select_description", "tsjsonc")}

This is the S3 method of the \code{\link[tsitter:ts_tree_select]{tsitter::ts_tree_select()}} generic,
for \link[=ts_tree_jsonc]{ts_tree_jsonc} objects.
}
\details{
\eval{tsitter:::doc_insert("tsitter::ts_tree_select_details", "tsjsonc")}
\eval{tsitter:::doc_insert("tsjsonc::ts_tree_select_examples", "tsjsonc")}
\eval{tsitter:::doc_extra()}
}
