% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-unserialize.R
\name{ts_tree_unserialize.ts_tree_jsonc}
\alias{ts_tree_unserialize.ts_tree_jsonc}
\title{Unserialize selected elements from a ts_tree_jsonc object}
\usage{
\method{ts_tree_unserialize}{ts_tree_jsonc}(tree)
}
\arguments{
\item{tree}{ts_tree_jsonc object.}
}
\value{
List of R objects, each the unserialization of a selected element
in tsjsonc.
}
\description{
Uses \code{\link[tsitter:ts_tree_unserialize]{tsitter::ts_tree_unserialize()}} on the selected elements.
}
\details{
If \code{json} does not have a selection, then all of it is unserialized.
If \code{json} has an empty selection, then an empty list is returned.
}
\examples{
library(tsitter)
json <- ts_parse_jsonc(ts_serialize_jsonc(list(
  a = list(a1 = list(1,2,3), a2 = "string"),
  b = list(4, 5, 6),
  c = list(c1 = list("a", "b"))
)))
json |> ts_tree_select(c("b", "c")) |> ts_tree_unserialize()
}
\seealso{
\code{\link[tsitter:ts_tree_unserialize]{tsitter::ts_tree_unserialize()}} to unserialize a JSON document from a
file or string. \code{\link[=ts_serialize_jsonc]{ts_serialize_jsonc()}} to create JSON from R objects.
}
\keyword{internal}
