% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-update.R
\name{ts_tree_update.ts_tree_jsonc}
\alias{ts_tree_update.ts_tree_jsonc}
\title{Replace selected JSON elements with a new element}
\usage{
\method{ts_tree_update}{ts_tree_jsonc}(tree, new, options = NULL, ...)
}
\arguments{
\item{tree}{ts_tree_jsonc object.}

\item{new}{R object that will be serialized to JSON (using
\code{\link[=ts_serialize_jsonc]{ts_serialize_jsonc()}}) and inserted in place of the selected JSON
elements.}

\item{options}{Named list of formatting options, see
\link[=tsjsonc_options]{tsjsonc options}.}

\item{...}{Reserved for future use.}
}
\value{
The updated ts_tree_jsonc object
}
\description{
Replace all selected elements with a new element. If \code{tree} has no
selection then the whole document is replaced.
}
\examples{
library(tsitter)
tree <- ts_parse_jsonc("{ \"a\": true, \"b\": [1, 2, 3] }")
tree

tree |> ts_tree_select("a") |> ts_tree_update(list("new", "element"))
}
\keyword{internal}
