% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-json.R
\name{tsjson-brackets}
\alias{tsjson-brackets}
\alias{[.tsjson}
\title{Convert a tsjson object to a data frame}
\usage{
\method{[}{tsjson}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{tsjson object.}

\item{i, j}{indices.}

\item{drop}{Ignored.}
}
\value{
A data frame with columns:
\code{id}, \code{parent}, \code{field_name}, \code{type}, \code{code}, \code{start_byte}, \code{end_byte}, \code{start_row}, \code{start_column}, \code{end_row}, \code{end_column}, \code{children}, \code{tws}.
}
\description{
Create a data frame for the syntax tree of a JSON document, by indexing
a tsjson object with single brackets. This is occasionally useful for
exploration and debugging.
}
\examples{
json <- load_json(text = serialize_json(list(
  a = list(a1 = list(1,2,3), a2 = "string"),
  b = list(4, 5, 6),
  c = list(c1 = list("a", "b"))
)))

json

json[]
}
\seealso{
Other JSON debugging tools: \code{\link[=sexpr_json]{sexpr_json()}}, \code{\link[=syntax_tree_json]{syntax_tree_json()}},
\code{\link[=query_json]{query_json()}}. \code{\link[=load_json]{load_json()}} for creating tsjson objects.
}
