% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{tsjson_options}
\alias{tsjson_options}
\title{tsjson options}
\description{
Options that control the behavior of tsjson functions.
}
\details{
\itemize{
\item \code{allow_empty_content}: logical, whether to allow empty JSON documents.
Default is \code{TRUE}.
\item \code{allow_comments}: logical, whether to allow comments in JSON documents.
Default is \code{TRUE}.
}
}
