% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unserialize-json.R
\name{unserialize_selected}
\alias{unserialize_selected}
\title{Unserialize selected elements from a tsjson object}
\usage{
unserialize_selected(json)
}
\arguments{
\item{json}{tsjson object.}
}
\value{
List of R objects, each the unserialization of a selected element
in tsjson. If \code{json} does not have a selection, then all of it is
unserialized. If \code{json} has an empty selectedion, then an empty list
is returned.
}
\description{
TODO
}
