% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-selection.R
\name{update_selected}
\alias{update_selected}
\title{Replace selected JSON elements with a new element}
\usage{
update_selected(json, new, format = c("pretty", "compact", "oneline"))
}
\arguments{
\item{json}{tsjson object.}

\item{new}{R object that will be serialized to JSON (using
\code{\link[=serialize_json]{serialize_json()}}) and inserted in place of the selected JSON
elements.}

\item{format}{How to format \code{new}. See \code{\link[=format_selected]{format_selected()}}.}
}
\value{
The updated tsjson object
}
\description{
Replace all selected elements with a new element. If \code{json} has no
selection then the whole document is replaced. If \code{json} has an empty
selection, then nothing happens.
}
\examples{
json <- load_json(text = "{ \"a\": true, \"b\": [1, 2, 3] }")
json

json |> select("a") |> update_selected(list("new", "element"))
}
