% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format.tsjson}
\alias{format.tsjson}
\title{Format a tsjson object}
\usage{
\method{format}{tsjson}(x, n = 10, ...)
}
\arguments{
\item{x}{tsjson object.}

\item{n}{Number of lines, or number of selections to print.}

\item{...}{Ignored.}
}
\value{
Character vector.
}
\description{
Format a tsjson object for printing.
}
\details{
This is the engine of \code{\link[=print.tsjson]{print.tsjson()}}, possibly useful to obtain a
printed representation without doing the actual printing.
}
\examples{
json <- load_json(text = serialize_json(list(
  a = list(a1 = list(1,2,3), a2 = "string"),
  b = list(4, 5, 6),
  c = list(c1 = list("a", "b"))
)))

json
}
