% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-json.R
\name{format_selected}
\alias{format_selected}
\title{Format the selected JSON elements}
\usage{
format_selected(json, format = c("pretty", "oneline", "compact"))
}
\arguments{
\item{json}{tsjson object.}

\item{format}{How to format \code{new}. See \code{\link[=format_selected]{format_selected()}}.}
}
\value{
The updated tsjson object.
}
\description{
TODO
}
