% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-json.R
\name{load_json}
\alias{load_json}
\title{Parse a JSON file or string into a tsjson object}
\usage{
load_json(file = NULL, text = NULL)
}
\arguments{
\item{file}{Path of a JSON file. Use either \code{file} or \code{text}.}

\item{text}{JSON string. Use either \code{file} or \code{text}.}
}
\value{
A tsjson object.
}
\description{
TODO
}
