% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tsjson}
\alias{print.tsjson}
\title{Print a tsjson object}
\usage{
\method{print}{tsjson}(x, n = 10, ...)
}
\arguments{
\item{x}{tsjson object.}

\item{n}{Number of lines, or number of selections to print.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Calls \code{\link[=format.tsjson]{format.tsjson()}} to format the tsjson object, writes the
formatted object to the standard output, and returns the original object
invisibly.
}
