% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select-set}
\alias{select-set}
\alias{select<-}
\alias{[[<-.tsjson}
\alias{deleted}
\title{Update selected elements in a tsjson object}
\usage{
select(json, ...) <- value

\method{[[}{tsjson}(x, i) <- value
}
\arguments{
\item{value}{New value. Will be serialized to JSON with \code{\link[=serialize_json]{serialize_json()}}.}

\item{x, json}{tsjson object. Create a tsjson object with \code{\link[=load_json]{load_json()}}.}

\item{i, ...}{Selectors, see \code{\link[=select]{select()}}.}
}
\value{
The updated tsjson object.

\code{\link[=deleted]{deleted()}} returns a marker object to be used at the right
hand side of the \code{\link[=select<-]{select<-()}} or the double bracket replacement
functions, see examples below.
}
\description{
Update the selected elements of a JSON document, using the replacement
function syntax.
}
\details{
Technically \code{\link[=select<-]{select<-()}} is equivalent to \code{\link[=select_refine]{select_refine()}} plus
\code{\link[=update_selected]{update_selected()}}. In case when \code{value} is

\code{\link[=deleted]{deleted()}} is a special marker to delete elements from a tsjson object
with \code{\link[=select<-]{select<-()}} or the double bracket operator.
}
\examples{
json <- load_json(text = "{}")

json <- json |> select("r", "editor.formatOnSave") |> update_selected(TRUE)
json

json <- json |> select("r", "editor.formatOnSave") |> delete_selected()
json

# Insert an array
json <- json |> select("foo") |> update_selected(1:3)
json

# Update the array at location 2
json |> select("foo", 2) |> update_selected(0)

# Insert at location 2
json |> select("foo") |> insert_into_selected(0, at = 2)

# Insert at the end of the array with `Inf` as `at`
json |> select("foo") |> insert_into_selected(0, at = Inf)

# Only the modified elements are reformatted
json <- load_json(text = '{"foo":[1,2],\n"bar":1}')
json |> select("foo") |> insert_into_selected(0, at = Inf)

# You can control how those elements are formatted
json |> select("foo") |>
  insert_into_selected(0, at = Inf, options = list(indent_width = 2))

# Using `deleted()` to delete elements
json <- load_json(text = serialize_json(list(
  a = list(a1 = list(1,2,3), a2 = "string"),
  b = list(4, 5, 6),
  c = list(c1 = list("a", "b"))
)))

select(json, list("a", "a1")) <- deleted()
json

json[[list("a", "a2")]] <- deleted()
json
}
\seealso{
Save the updated tjson object to a file with \code{\link[=save_json]{save_json()}}.
}
