% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select<-}
\alias{select<-}
\alias{[[<-.tsjson}
\title{Update selected elements in a tsjson object}
\usage{
select(json, ...) <- value

\method{[[}{tsjson}(x, i) <- value
}
\arguments{
\item{json}{tsjson object. Create a tsjson object with \code{\link[=load_json]{load_json()}}.}

\item{...}{Selectors, see \code{\link[=select]{select()}}.}

\item{value}{New value. Will be serialized to JSON with \code{\link[=serialize_json]{serialize_json()}}.}
}
\value{
The updated tsjson object.
}
\description{
TODO
}
\seealso{
Save the updated tjson object to a file with \code{\link[=save_json]{save_json()}}.
}
