% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sitter.R
\name{sexpr_json}
\alias{sexpr_json}
\title{Show the syntax tree structure of a JSON file or string}
\usage{
sexpr_json(file = NULL, text = NULL, ranges = NULL)
}
\arguments{
\item{file}{Path of a JSON file. Use either \code{file} or \code{text}.}

\item{text}{JSON string. Use either \code{file} or \code{text}.}

\item{ranges}{Optionally a list of integer pairs defining ranges within
\code{file} or \code{text} to parse.}
}
\description{
Show the syntax tree structure of a JSON file or string
}
\examples{
sexpr_json(text = "{ \"a\": true, \"b\": [1, 2, 3] }")
}
