% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsjson-package.R
\docType{package}
\name{tsjson-package}
\alias{tsjson}
\alias{tsjson-package}
\title{tsjson: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
