% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsjson-package.R
\docType{package}
\name{tsjson-package}
\alias{tsjson}
\alias{tsjson-package}
\title{tsjson: Edit JSON Files}
\description{
Extract and manipulate parts of JSON files without touching the formatting and comments in other parts.
}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
  \item  tree-sitter-json authors [contributor]
}

}
\keyword{internal}
