% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{tsjson_options}
\alias{tsjson_options}
\title{tsjson options}
\description{
Options that control the behavior of tsjson functions.
}
\details{
\subsection{Parsing options:}{
\itemize{
\item \code{allow_empty_content}: logical, whether to allow empty JSON documents.
Default is \code{TRUE}.
\item \code{allow_comments}: logical, whether to allow comments in JSON documents.
Default is \code{TRUE}.
}
}

\subsection{Formatting options:}{
\itemize{
\item \code{indent_width}: integer, the number of spaces to use for indentation
when \code{indent_style} is \code{"space"}. Default is \code{4}.
\item \code{indent_style}: string, either \code{"space"} or \code{"tab"}, the type of
indentation to use. Default is \code{"space"}.
}
}
}
